main navigation components extracted
